[redux-dbg] Add redux remotedev server
Summary: This adds the remotedev server to dependencies and adds the script that starts it on `yarn dev` in native alongside with metro. This server is necessary to remotely debug redux when using hermes.

Test Plan: Run `yarn dev` and check if both of these ports are busy: `8081`(metro), `8000`(remotedev server)

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, KatPo, zrebcu411, Adrian, atul, subnub

Differential Revision: https://phabricator.ashoat.com/D936