[redux-dbg] Add `remoteDevServerURLPrefixAsync`
Summary: Adds a function to fetch the information whether the app is running on emulator or not. It uses [this function](https://github.com/react-native-device-info/react-native-device-info/tree/7f435dceeb52b5adbac55af2d48f1a257fc1636e#isemulator) from `react-native-device-info`. The function `isEmulator` is asynchronous, so is the added one here - `remoteDevServerURLPrefixAsync`

Test Plan: Try to run the app on emulator and on a physical device and see a different output for both cases.

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, KatPo, zrebcu411, Adrian, atul, subnub

Differential Revision: https://phabricator.ashoat.com/D937