[lib] Block creating sidebars from failed messages
Summary: We can simply check if message server id is present. If it is, the message was created successfully and if not, then it's only a local message and creating sidebar from it should be blocked.

Test Plan:
Send failed text and multimedia message on web and native and check if create sidebar option was not displayed.
Check again each of the cases with successful message and check if the option was visible.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, zrebcu411, Adrian, atul, subnub

Differential Revision: https://phabricator.ashoat.com/D935