[web] Add active item to currently opened tab only
Summary: When opening thread user is not a member of, we want to add thread to currently opened tab as active thread item.

Test Plan: Open non-member thread and change tab - notice thread in both tabs. With this diff notice thread added only to tab that was active when thread was opened

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, zrebcu411, Adrian, atul, subnub

Differential Revision: https://phabricator.ashoat.com/D945