[server] Don't include KNOW_OF in SIDEBAR roles
Summary: In combination with D955, this makes it so membership in a sidebar is no longer a sufficient condition to have `KNOW_OF` with that sidebar. Losing `KNOW_OF` will delete the membership, meaning that the `KNOW_OF` has to come from the parent for you to be able to stay a member of a sidebar.

Test Plan: Create a new sidebar in a non-`PRIVATE` thread (so the parent thread already has the right roles). Then try to add somebody outside the group to the sidebar (after commenting out the code on both native and server that would normally prevent this) and confirm it doesn't work

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1020