[web] Display sidebar parent as active thread item when user is not member of this thread
Summary:
When user is not member of a thread T, we will not display T in chat thread list unless it's active - non-member threads are filtered out from chatListData. We want to display thread T is list in one more case: when it's sidebar (S) becomes active. S is not top-level thread and can't be displayed in thread list without its parent T. But when S becomes active, T is no longer included in thread list.
We need to have a way to keep T in thread list when it's child S becomes active.
I introduced activeSidebarParentThreadItem, which will be used to get activeChatThreadItem parent. It'll be added to chatListData if active item is a sidebar, and does not have parent in thread list already.

Test Plan: Open sidebar from non-member thread and check if active thread is displayed correctly, with sidebar displayed below. Tested together with D945 to make sure threads are added to one tab only

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, zrebcu411, Adrian, atul, subnub

Differential Revision: https://phabricator.ashoat.com/D956