[server] Separate out global and app-specific URL config
Summary: Note: this is a breaking change. We used to have everything in `server/facts/url.json`, but now some of it is in `server/facts/app_url.json`. Everyone on the team will need to go ahead and create the latter file, following the updated instructions in `dev_environment.md` (see below).

Test Plan: Make sure server still returns a working website

Reviewers: atul, palys-swm

Reviewed By: atul

Subscribers: KatPo, zrebcu411, Adrian, subnub

Differential Revision: https://phabricator.ashoat.com/D966