[server] Avoid setting role to -1 when permissions are non-null
Summary: The other two changeset-generating functions handle this correctly, but `recalculateThreadPermissions` does not. The `role` variable here is only used if the permissions are non-null, and since we are trying to maintain a principle that null `permissions` occurs if and only if `role` is -1, we need to make sure the role variable here is always non-negative.

Test Plan: I'm going to submit a whole bunch of permissions diffs and will aim to test them together later. This one should be pretty clear, though

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1005