[server] Narrow hasParentPermission check in updateThread
Summary: We only need to check this condition unless the parent thread ID has changed, or the thread has switched from being a sidebar to a non-sidebar (or vice versa). Neither `forceUpdateRoot` nor a thread type change between `CHAT_NESTED_OPEN` and `CHAT_SECRET` is sufficient.

Test Plan: I noticed this issue when I was trying to change a test thread type. The UI allowed it but the server did not. I concluded that the UI was correct in this case, and made sure that after this diff the change went through.

Reviewers: KatPo

Reviewed By: KatPo

Subscribers: zrebcu411, palys-swm, Adrian, atul, subnub

Differential Revision: https://phabricator.ashoat.com/D974