[web] Find tooltip position for tooltip with multiple items
Summary:
With tooltip menu introduction, we need to change how findTooltipPosition works.
1. We need to get correct value of tooltipWidth, which will be calculated based on longest text in tooltip menu
2. We need to calculate tooltipHeight based on number of items in tooltip menu
3. We need to make sure left/right position will be used only for single element tooltips

Test Plan: Check if tooltips work as before. Add string to `tooltipTexts`, check if longest is detected correctly, check if height of tooltip is correct. Check if left and right positions can be used only when there is one item

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D990