[server] Use getThreadTypeParentRequirement in createThread
Summary:
I first introduced this function in D784 to use with `updateThread`.

The logic here should be unchanged, except that we will now throw when trying to create a `PRIVATE` thread with a parent. This was not possible from the API before since `PRIVATE` threads can only be created programmatically (blocked by `newThreadRequestInputValidator`), but it's still an improvement.

Test Plan: Flow, careful reading

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1002