[server] Only return member sessions in fetchActiveSessionsForThread
Summary: This function is used to determine which users to send an update about a calendar entry to. The code in `entry-reducer.js` will filter out entries for threads the user is not a member of, so there's no point sending thse entries to users that aren't members.

Test Plan: Create a subthread that a test user can see but isn't a member of. Before this diff, the Redux monitor showed new entries in that subthread being delivered to the test user, but after this diff they aren't

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1015