[lib] Don't include viewer's username in threadSearchText
Summary:
I was searching for my own username and hoping to pull up my `PRIVATE` thread, but got a bunch of `PERSONAL` threads above it.

We should probably prioritize the `PRIVATE` thread at the top (or at least equal to `PERSONAL`). But really the issue here I think is that we are including the viewer's username in `threadSearchText`. Since the viewer is a member of every thread in the thread list, this means searching for your own username will match every thread, which isn't very helpful.

Test Plan: Test searching the viewer's username and make sure it doesn't everything indiscriminately

Reviewers: palys-swm, KatPo

Reviewed By: palys-swm, KatPo

Subscribers: Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1016