[server] Throw exception in changeRole if intent not handled
Summary: This should always get caught earlier, but in case it doesn't, we should error in `thread-permission-updaters.js`.

Test Plan: Create a new sidebar in a non-`PRIVATE` thread (so the parent thread already has the right roles). Then try to add somebody outside the group to the sidebar (after commenting out the code on both native and server that would normally prevent this) and confirm that it throws this error

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1028