Remove duplicated entries data from joinThread response
Summary:
In the near future, we want to call `joinThread` from `createThread`. In order to do so, we need to simplify `joinThread` and remove fetching that is duplicated there - entries and messages.
This diff removes calendar-fetching logic from `joinThread`

Test Plan: Join thread and check if there are no errors. Check if entries are added to entryStore, check if data is not duplicated

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1044