[web] Reassign drafts from pending to realized threads
Summary:
We could do it in componentDidUpdate but that would result in one render without the draft.
Another alternative was to translate thread ids in inputStateSelector, but then the readability would suffer. Also, we want to run the reassignment before componentDidiUpdate code (so that messages get correct threadID, which would be hard with this approach).

Test Plan: Type some text in pending thread input bar, create realized thread on a mobile and check if draft is still there.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1049