[web] Copy thread creation logic from native
Summary:
In D844 we introduced thread creation logic to input state container on native and significantly changed it in D1034.
We created a [[ https://www.notion.so/commapp/Merge-native-and-web-input-state-container-831aa218b46b48f1bcdde2108c97351c | Notion issue ]] to merge native and web input state containers. I think that extracting common parts should be done as a part of it, so it's fine to have this logic duplicated.

Test Plan: Flow. This function is not used, so just check if app launches. This was tested extensively in D1063.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1072