[lib] Allow log in to update checkpoint
Summary:
Right now, because log in actions (log in, reset password, register) can't update checkpoints, we send up 0 checkpoints on `INCREMENTAL_STATE_SYNC`.

The result of this is that the socket can't connect because the state sync times out trying to fetch literally every message this user has ever received.

Before D908 this issue was only affecting the updates checkpoint, which was bad but not bad enough to cause too many issues because we delete updates after they're fully delivered.

Test Plan: Make sure the socket can connect after log-in on a newly deployed app, and that the checkpoints are non-zero in the state sync code on the server

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1066