[native] Update urlPrefix on DEV after checking if device is emulator
Summary:
When deploying to a real device, right now we default the `urlPrefix` to assuming the device is an emulator (eg. `10.0.2.2` on Android, `localhost` on iOS).

This makes the real device unable to reach the Node dev server. Let's make sure we update the `urlPrefix` after determining if the device is an emulator.

Test Plan: Make sure we can connect to the dev server from an app that is newly deployed (eg. no previous Redux state) on a real device

Reviewers: karol-bisztyga

Reviewed By: karol-bisztyga

Subscribers: KatPo, palys-swm, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1067