[native] Silence errors about remote Redux dev server connection
Summary:
Currently, if you are using Android (rather, Hermes) we run `remotedev.connect` on the device to connect with the `remotedev` server running on your laptop on port 8043.

If this connection fails it gives a red box, which is very aggressive. I'd rather it just logged an error.

I investigated the underlying library `socketcluster-client` and found that it doesn't throw exceptions if there is a listener listening for `error` events.

Test Plan: Make sure I don't see a red box if I'm not running the `remotedev` server, and instead I see log messages for the errors

Reviewers: karol-bisztyga

Reviewed By: karol-bisztyga

Subscribers: KatPo, palys-swm, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1068