[native] Temporary for Android failing to load large Redux store
Summary:
Android wasn't persisting the Redux store when logged in with the "ashoat" user. I saw the "Row too big to fit into CursorWindow" error, so I Googled around and found this hack in a [GitHub issue](https://github.com/react-native-async-storage/async-storage/issues/537).

The long-term fix is for us to move the database stuff to SQLite. For completeness, here are some other fixes I considered:

1. [react-native-mmkv](https://github.com/mrousavy/react-native-mmkv) as storage for `redux-persist` (not confident it will handle huge pages better though)
2. [Separating reducers](https://github.com/rt2zz/redux-persist/issues/1265#issuecomment-742704826) into distinct `AsyncStorage` keys (not sure what the distribution of size by keys is, and might require a migration)
3. Upgrading `AsyncStorage` and investing [this](https://react-native-async-storage.github.io/async-storage/docs/advanced/db_size) (people on the issue said it didn't help)
4. [Moving](https://github.com/react-native-async-storage/async-storage/pull/528) off of SQLite storage onto Android Room storage (people on the issue said it didn't help)

I think using this hack for now is a good stopgap.

Test Plan: Make sure Android is loading the persisted store correctly when the app starts

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1069