[native] Use correct payload when dispatching `sendTextMessageActionTypes.failed` action
Summary:
In D844 we started dispatching `sendTextMessageActionTypes.failed` but the payload was incorrect. The issue doesn't manifest in any way during testing because in message-reducer we're only interested in `localID` part of the payload, which was also present when messageInfo was sent instead of an error.
There's one more part, which we should take a look at: the action has one more required parameter, `loadingInfo`. It's only used by `loading-reducer` and it handles undefined case well, so we can consider marking this property as optional in `redux-types`.

Test Plan: Create failed message and check using the debugger if payload contains an error. The payload should have an additional property `name` just like it has when we send text message using `dispatchActionPromise`

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1073