[lib] Include threadType in getPendingThreadID
Summary:
I noticed an issue where we would show a `PERSONAL` `createPendingThreadItem` in the search results for users for whom there was no existing `PERSONAL` thread, but when the thread was pressed it would turn into an existing non-`PERSONAL` thread.

This is definitely an edge case (you should usually have a `PERSONAL` thread before a non-`PERSONAL` thread with a user), but it's possible if somebody adds both of you to a `CHAT_SECRET` thread and then leaves.

I think the right solution is that we should include `threadType` in the pending thread ID, so that a pending thread ID of a particular type can only be replaced by a realized thread of the same type.

Test Plan: Make sure that pressing a `PERSONAL` `createPendingThreadItem` in the search results doesn't redirect to a realized non-`PERSONAL` thread

Reviewers: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1094