[lib] Update message timestamp when realized thread is detected
Summary:
This diff solves a bug where pending message is displayed before `SIDEBAR_SOURCE`, `CREATE_SIDEBAR` and `CREATE_THREAD` messages. It was happening because when a message is sent to a pending thread, it has a local timestamp which is older than a realized thread. To solve that, we need to increase a timestamp in two places:
- When we detect that a new thread is created, we should set the timestamp of messages which are already in the store. We can do that safely, because when these will be created on the server, their timestamps will be even greater.
- When we create a realized thread by sending a message. Even though this message is in the store during reassignment, the timestamp will be overridden when `sendXXXMessageActionTypes` action is dispatched using `dispatchActionPromise`.

Test Plan: Check sending text and multimedia message on web and text on native. Send a message to a pending thread and verify that the "jumping" no longer appear.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1109