[native] Set controls to visible when user backgrounds video
Summary: Currently, we pause video playback when the user backgrounds the app. However, the video controls may or may not be visible  when the user returns. Now, we'll set `controlsVisible` to `true` when the app is backgrounded, so the user always sees the play button when they return to the app/can pick up where they left off.

Test Plan: Tried backgrounding and returning before/after change with controls hidden/visible before backgrounding... and it worked as expected.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1110