[lib] Make sure that sidebar source is always displayed
Summary:
This diff fixes the following bug: when one user (A) opens a pending sidebar we include source message in the list. User A sends a message to the thread, but it fails. Then another user (B) creates the thread. From A's perspective, there's a delay between thread creation and messages update, so for some time we display a thread without sidebar source. One possible solution was to modify thread creation logic to send all the updates at once, but that seemed really complicated.
The solution is to include the source message even when a thread is realized. We have to make sure that we (1) don't include a duplicated message and (2) that the message doesn't have a link to the sidebar. To achieve the 1st property, we could check if sidebar contains SIDEBAR_SOURCE, but that's not enough. When some messages were pruned, the thread might have the source but not in this user's store. This case is easy, we can simply check if the start was reached and add the source only if that's the case. The 2nd property requires setting `threadCreatedFromMessage` property to undefined for the sidebars. We don't plan to introduce recursive sidebars in the near future, so it sounds ok.

Test Plan: Do what's in the summary and check if the behavior is correct. I'll include videos from before and after the bugfix.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1118