[lib] Correctly handle pending threads in started message actions reducer
Summary:
In D1100 I introduced creating a thread when local message is sent. The shape of the thread is incorrect, so I had to introduce correct thread creation. The bug started to be visible after D1118.
Additionally, the reducer logic for message start had some duplication regarding messageIDs, so I improved it.

Test Plan: Send failed multimedia message on native in pending sidebar and check if source message is displayed.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1131