[cpp] Copy-paste react-native internal Promise utils into separate file
Summary:
Sharing this to show one possible way I was able to get Android to compile `createPromiseAsJSIValue`.

Ultimately this sort of defers the problem of getting `folly` to compile on Android, so it's probably not a good long-term solution.

Test Plan: Make sure it compiles and runs on iOS and Android, and that drafts read / write correctly on both

Reviewers: karol-bisztyga, palys-swm

Reviewed By: karol-bisztyga

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1122