[cpp] Rename threadID column to key
Summary:
This is mostly an academic exercise... the column name hardly matters. But I was curious how we'll handle migrations going forward.

In this case, it's okay to repeatedly run this query without error handling at startup. In the future, we'll probably want to store some state about the database schema version somewhere, and migrate the schema based on the current version.

@karol-bisztyga, what do you think?

Test Plan: Compile and run before diff, save a draft, then compile and run with diff, and make sure same draft loads

Reviewers: karol-bisztyga, palys-swm

Reviewed By: karol-bisztyga, palys-swm

Subscribers: KatPo, Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1113